2.20110424BX100 - Added new sendmail.mc.mailman with provisions for Mailman - Added new sendmail.mc.majordomo with provisions for Majordomo - Modified glue/configs/sendmail/sendmail.mc to have no support for Mailman or Majordomo - Updated glue/scripts/initSendmail.sh with provisions for Mailman and Majordomo

+++ BlueOnyx/5106R/common/base-blueonyx.mod/glue/scripts/initSendmail.sh
@@ -13,7 +13,29 @@ if [ ! -f /etc/mail/aliases ]; then
   fi
   /usr/sbin/makemap hash /etc/mail/aliases.db < /etc/mail/aliases
 fi
-cp -p $CONFDIR/sendmail.mc /etc/mail/sendmail.mc
+
+# Handle Mailman presence:
+if [ -f /etc/rc.d/init.d/mailman ]; then
+	cp -p $CONFDIR/sendmail.mc.mailman /etc/mail/sendmail.mc
+else
+	touch /tmp/nolist.mailmain
+fi
+
+# Handle Majordomo presence:
+if [ -f /usr/local/majordomo/bin/approve ]; then
+	cp -p $CONFDIR/sendmail.mc.majordomo /etc/mail/sendmail.mc
+else
+	touch /tmp/nolist.majordomo
+fi
+
+# Handle case where we have neither and a stock Sendmail config:
+/bin/grep 'setup for BlueOnyx' /etc/mail/sendmail.mc > /dev/null 2>&1
+if [ $? = 0 ]; then 
+	if [ ! -f /tmp/nolist.mailmain ] && [ ! -f /tmp/nolist.majordomo ];then
+		cp -p $CONFDIR/sendmail.mc /etc/mail/sendmail.mc
+	fi
+fi
+
 cp -p $CONFDIR/popauth.m4 /usr/share/sendmail-cf/hack/popauth.m4
 
 if [ -f /usr/sausalito/constructor/solarspeed/av_spam/aa_initial_inst.pl ];then
